moved safesnap composable to plugin dir
