+++ src/plugins/safeSnap/components/Form/TransactionBatch.vue
@@ -1,6 +1,6 @@
 <script>
 import { clone } from '@snapshot-labs/snapshot.js/src/utils';
-import { useSafesnap } from '@/composables/useSafesnap';
+import { useSafesnap } from '@/plugins/safeSnap/composables/useSafesnap';
 import chevronIcon from '@/assets/icons/chevron.svg';
 import { createBatch, ERC20_ABI, ERC721_ABI } from '../../index';
 import { formatEther } from '@ethersproject/units';
